[dispatch] allow static command handlers to be invoked without no-args constructor
Previously, there was assumed to be a no-args constructor to construct an instance of the object to be invoked.  If there is no no-args constructor this would not be able to create an instance.  While not creating an instance is acceptable behavior in that case, we still should support invocation of static command handlers that do not require an instance at all.  This change should support static invocation and not change anything about previous usage.