[dispatch] fix single / double prefix throwing errors and giving poor behavior
Assuming prefix = !

"!" would throw an Out Of Bounds Exception.

"!!" would replace both instances with nothing, and would also throw an OOB Exception

This is fixed by using String substringing instead of replacing anything.  It should be a bit less fragile.  Prefixing on empty string was also fixed as that would be equivalent to "!help" in a roundabout way.